+++ libraries/AD56X8/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.2] - 2023-01-09
+- update GitHub actions
+- update license
+- add wave example (triangle)
+- add bool setPercentage(channel, percentage)
+- add float getPercentage(channel)
+
+
 ## [0.1.1] - 2022-12-01
 - initial release
 - add changelog.md
@@ -14,7 +22,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - improve unit tests
 - clean up code
 - update readme.md
-
 
 ## [0.1.0] - 2022-07-28
 - rework (add from DAC855x + MCP_DAC)